Add files via upload
It's a voice bot. it take voice file and detect it's language and print it after that convert it into text and translate into english and print it also match the intent with already present word and if it is match then genrate the response text otherwiese go for live agent.